+++ CHANGELOG.md
@@ -16,6 +16,11 @@ The format is based on [Keep a Changelog].
 
 ## [UNRELEASED]
 
+### Added
+
+- `IBMQJob` now has three new methods: `done()`, `running()`, and
+  `cancelled()`. The methods are used to indicate the job status. (\#494)
+
 ### Changed
 
 - The Exception hierarchy has been refined with more specialized classes.